Moved webdriver_path and binary_path from config to the variants file.
