+++ src/lqc/config/config.py
@@ -1,6 +1,4 @@
-import subprocess
 import json
-
 
 DEFAULT_STYLE_WEIGHT = 10
 DEFAULT_STYLE_VALUE_WEIGHT = 10
@@ -32,11 +30,6 @@ class Config:
             paths = config.get("paths", {})
             cls.__instance.path_bug_reports_dir = paths.get("bug-reports-directory", "./bug_reports")
             cls.__instance.path_tmp_files_dir = paths.get("tmp-files-directory", "./tmp_generated_files")
-            cls.__instance.path_chrome_driver = paths.get("chrome-webdriver", None) or Config.detectDriverPath("chromedriver", "chrome-webdriver")
-            cls.__instance.path_chrome_binary = paths.get("chrome", None)
-            cls.__instance.path_firefox_driver = paths.get("firefox-webdriver", None) or Config.detectDriverPath("geckodriver", "firefox-webdriver")
-            cls.__instance.path_firefox_binary = paths.get("firefox", None)
-            cls.__instance.path_safari_driver = paths.get("safari-webdriver", None) or Config.detectDriverPath("safaridriver", "safari-webdriver")
 
         elif cls.__instance == None:
             raise RuntimeError("Config must be initialized before use")
@@ -62,29 +55,4 @@ class Config:
 
     def getTmpFilesDirectory(self):
         return self.path_tmp_files_dir
-
-    def getChromeDriverPath(self):
-        return self.path_chrome_driver
-
-    def getChromeBinaryPath(self):
-        return self.path_chrome_binary
-
-    def getFirefoxDriverPath(self):
-        return self.path_firefox_driver
-
-    def getFirefoxBinaryPath(self):
-        return self.path_firefox_binary
-
-    def getSafariDriverPath(self):
-        return self.path_safari_driver
-    
-    @staticmethod
-    def detectDriverPath(driver, config_name):
-        status, driver_path = subprocess.getstatusoutput(f"which {driver}")
-        if status == 0:
-            print(f"Warning: No {config_name} path in config. Using {driver_path}")
-            return driver_path
-        else:
-            print(f"Warning: No {config_name} Found. You may need to install {driver} or put the path in the config.")
-            return None
